requirements.txt
Django>=4.0
numpy
opencv-python
dlib
insightface
onnxruntime
Pillow>=9.0
requests